requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiohttp==3.9.1
certifi>=2023.7.22

